fix: address issue with assistant menu in editor
